[BugFix] Make pipeline_driver's total_timer exclude the prepare time (#24899)
Fixes #24534

`PipelineDriver::prepre` is executed in a serial way. And some of the
timers are triggered at prepare stage. And the total time may include
other driver's prepare time usage, which further results in an abnomal
ScheduleTime in profile if other dirver's prepare stage persists for a
long while.

## What type of PR is this:
- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [x] 3.0
  - [x] 2.5
  - [x] 2.4
  - [x] 2.3

---------

Signed-off-by: liuyehcf <1559500551@qq.com>