[BugFix] reset index id in MaterializedIndexMeta when restore (#25154) (#25155)
Fixes #25154
Restore will reset index id in MaterializedIndex, but currently does not reset the corresponding index id in MaterializedIndexMeta. This will cause the problem that when the table has some rollup index, and we backup and restore it. It will get NPE when we query it. Because, we want to rewrite the query using materialized view but we have no correct id information in MaterializedIndexMeta.

(cherry picked from commit 8426916bf2cec57e8827649b88cdac2602383638)