[BugFix] Fix date type as partition value when inserting iceberg table (#26760) (#26929)
Fixes #issue
We need to get date string like "YYYY-MM-dd" from Datum. But the value
in the DateColumn is JuLian of DateValue. It's an int_32 type. So we
need to transform it to date string. Additionally, we should restrict
unsupported data types from being used as partition columns.
port: https://github.com/StarRocks/starrocks/pull/26760

Signed-off-by: stephen <stephen5217@163.com>